move ProfileCard in ProfileContent and change type of chooseImage state
