+++ src/components/ProfileContent/ProfileContent.tsx
@@ -18,28 +18,32 @@ import { useAppDispatch } from '../../redux/store';
 import { useAppSelector } from '../../redux/store';
 import { EditUserDataType } from '../../interfaces/User';
 import { toError } from '../../utils/error';
+import ProfileCard from '../ProfileCard/ProfileCard';
 
-type ProfileContentProps = {
-  isProfileInEditMode: boolean;
-  cancelEditClick: (prop: boolean) => void;
-  setChooseImage: (prop: File) => void;
-  setLoading: (prop: string) => void;
-  chooseImage: File;
-};
-let initialImageState: File;
-
-const ProfileContent = (props: ProfileContentProps) => {
+const ProfileContent = () => {
   const dispatch = useAppDispatch();
   const user = useAppSelector((state) => state.user);
   const [errorMsg, setErrorMsg] = useState<string>('');
   const [loadingRequest, setLoadingRequest] = useState(false);
+  const [isProfileInEditMode, setIsProfileInEditMode] = useState(true);
+  const [chooseImage, setChooseImage] = useState<File | undefined>(undefined);
+  const [instantPhoto, setInstantPhoto] = useState<string>('');
+  const [loading, setLoading] = useState('');
+
+  const changeProfileEditMode = (isInEditMode: boolean) => {
+    if (isInEditMode) {
+      setChooseImage(undefined);
+      setInstantPhoto('');
+    }
+    setIsProfileInEditMode(isInEditMode);
+  };
   async function dispatchEditUserRequest(userData: EditUserDataType) {
     try {
       setLoadingRequest(true);
       await dispatch(editUser(userData)).then(() => {
         setLoadingRequest(false);
       });
-      props.cancelEditClick(true);
+      changeProfileEditMode(true);
     } catch (error) {
       const err = toError(error);
       const errMessage = validateEdit(err);
@@ -71,7 +75,7 @@ const ProfileContent = (props: ProfileContentProps) => {
       lastNameRef.current.value = user.lastName || '';
       birthdayRef.current.value = user.birthday || '';
     }
-    props.cancelEditClick(true);
+    changeProfileEditMode(true);
   };
 
   const updateProfileFieldsDb = async () => {
@@ -95,7 +99,7 @@ const ProfileContent = (props: ProfileContentProps) => {
       userData.newFields.firstName = firstNameInputValue;
       userData.newFields.lastName = lastNameInputValue;
 
-      props.setChooseImage(initialImageState);
+      setChooseImage(undefined);
 
       if (jobRoleInputValue || user.jobRole) {
         userData.newFields.jobRole = jobRoleInputValue;
@@ -110,17 +114,17 @@ const ProfileContent = (props: ProfileContentProps) => {
       if (errMessage) {
         return;
       }
-      if (props.chooseImage) {
+      if (chooseImage) {
         const storageRef = ref(storage, user.userId);
-        const uploadImage = uploadBytesResumable(storageRef, props.chooseImage);
-        props.setLoading('uploading is...');
+        const uploadImage = uploadBytesResumable(storageRef, chooseImage);
+        setLoading('uploading is...');
         uploadImage.on(
           'state_changed',
           async (snapshot) => {
             const progress =
               (snapshot.bytesTransferred / snapshot.totalBytes) * 100;
             if (progress === 100) {
-              props.setLoading('uploading is ' + progress + '% done');
+              setLoading('uploading is ' + progress + '% done');
             }
           },
           (error) => {
@@ -144,153 +148,174 @@ const ProfileContent = (props: ProfileContentProps) => {
   };
 
   return (
-    <Container className="profile-content-container">
-      <Row className="first-row">
-        <Card className="card-content card-box">
-          <Card.Body className="card-content__body card-box__body">
-            <Card.Text className="card-content__text">
-              Some advice text.Some advice text.Some advice text.Some advice
-              text.Some advice text.Some advice text.Some advice text.Some
-              advice text.
-            </Card.Text>
-            <Card.Title className="card-content__title">Profile</Card.Title>
-          </Card.Body>
-        </Card>
-        <Col sm>
-          <Row>
-            <Col>First Name</Col>
-            <Col>
-              <InputField
-                className="profile-form"
-                reference={firstNameRef}
-                placeholder=""
-                defaultValue={user.firstName || ''}
-                readOnlyValue={props.isProfileInEditMode}
-              />
-            </Col>
-          </Row>
-          <Row>
-            <Col>Country</Col>
-            <Col>
-              {' '}
-              <InputField
-                className={`profile-form ${
-                  !user.country && 'profile-form--empty'
-                }`}
-                reference={countryRef}
-                placeholder={!props.isProfileInEditMode ? '' : 'your country'}
-                defaultValue={user.country || ''}
-                readOnlyValue={props.isProfileInEditMode}
-              />
-            </Col>
-          </Row>
-          <Row>
-            <Col>Job Role</Col>
-            <Col>
-              {' '}
-              <InputField
-                className={`profile-form ${
-                  !user.jobRole && 'profile-form--empty'
-                }`}
-                reference={jobRoleRef}
-                placeholder={!props.isProfileInEditMode ? '' : 'your job'}
-                defaultValue={user.jobRole || ''}
-                readOnlyValue={props.isProfileInEditMode}
-              />
-            </Col>
-          </Row>
-        </Col>
-        <Col sm>
-          <Row>
-            <Col>Last Name</Col>
-            <Col>
-              {' '}
-              <InputField
-                className="profile-form"
-                reference={lastNameRef}
-                placeholder=""
-                defaultValue={user.lastName || ''}
-                readOnlyValue={props.isProfileInEditMode}
-              />
-            </Col>
-          </Row>
-          <Row>
-            <Col>Birthday</Col>
-            <Col>
-              {' '}
-              <InputField
-                className={`profile-form ${
-                  !user.birthday && 'profile-form--empty'
-                }`}
-                reference={birthdayRef}
-                placeholder={!props.isProfileInEditMode ? '' : 'your birthday'}
-                defaultValue={user.birthday || ''}
-                readOnlyValue={props.isProfileInEditMode}
-              />
-            </Col>
-          </Row>
-          <Row>
-            <Col>Email</Col>
-            <Col className="first-row__email-column">{user.email}</Col>
-          </Row>
-        </Col>
-
-        {!props.isProfileInEditMode && (
-          <Row className="first-row__buttons">
-            {loadingRequest && (
+    <Container fluid className="profile-container">
+      <Container className="profile-container__profile-card">
+        <ProfileCard
+          activeEditClick={changeProfileEditMode}
+          isProfileInEditMode={isProfileInEditMode}
+          setChooseImage={setChooseImage}
+          setInstantPhoto={setInstantPhoto}
+          instantPhoto={instantPhoto}
+          setLoading={setLoading}
+          loading={loading}
+        />
+      </Container>
+      <Container className="profile-container__profile-content">
+        <Container className="profile-content-container">
+          <Row className="first-row">
+            <Card className="card-content card-box">
+              <Card.Body className="card-content__body card-box__body">
+                <Card.Text className="card-content__text">
+                  Some advice text.Some advice text.Some advice text.Some advice
+                  text.Some advice text.Some advice text.Some advice text.Some
+                  advice text.
+                </Card.Text>
+                <Card.Title className="card-content__title">Profile</Card.Title>
+              </Card.Body>
+            </Card>
+            <Col sm>
               <Row>
-                <Col colSpan={8} className="first-row__center-spinner">
-                  <Spinner animation="grow" className="bg-primary" />
+                <Col>First Name</Col>
+                <Col>
+                  <InputField
+                    className="profile-form"
+                    reference={firstNameRef}
+                    placeholder=""
+                    defaultValue={user.firstName || ''}
+                    readOnlyValue={isProfileInEditMode}
+                  />
                 </Col>
               </Row>
-            )}
-            {errorMsg && (
-              <div className="first-row__buttons--error error-message">
-                {errorMsg}
-              </div>
-            )}
+              <Row>
+                <Col>Country</Col>
+                <Col>
+                  {' '}
+                  <InputField
+                    className={`profile-form ${
+                      !user.country && 'profile-form--empty'
+                    }`}
+                    reference={countryRef}
+                    placeholder={!isProfileInEditMode ? '' : 'your country'}
+                    defaultValue={user.country || ''}
+                    readOnlyValue={isProfileInEditMode}
+                  />
+                </Col>
+              </Row>
+              <Row>
+                <Col>Job Role</Col>
+                <Col>
+                  {' '}
+                  <InputField
+                    className={`profile-form ${
+                      !user.jobRole && 'profile-form--empty'
+                    }`}
+                    reference={jobRoleRef}
+                    placeholder={!isProfileInEditMode ? '' : 'your job'}
+                    defaultValue={user.jobRole || ''}
+                    readOnlyValue={isProfileInEditMode}
+                  />
+                </Col>
+              </Row>
+            </Col>
+            <Col sm>
+              <Row>
+                <Col>Last Name</Col>
+                <Col>
+                  {' '}
+                  <InputField
+                    className="profile-form"
+                    reference={lastNameRef}
+                    placeholder=""
+                    defaultValue={user.lastName || ''}
+                    readOnlyValue={isProfileInEditMode}
+                  />
+                </Col>
+              </Row>
+              <Row>
+                <Col>Birthday</Col>
+                <Col>
+                  {' '}
+                  <InputField
+                    className={`profile-form ${
+                      !user.birthday && 'profile-form--empty'
+                    }`}
+                    reference={birthdayRef}
+                    placeholder={!isProfileInEditMode ? '' : 'your birthday'}
+                    defaultValue={user.birthday || ''}
+                    readOnlyValue={isProfileInEditMode}
+                  />
+                </Col>
+              </Row>
+              <Row>
+                <Col>Email</Col>
+                <Col className="first-row__email-column">{user.email}</Col>
+              </Row>
+            </Col>
 
-            <Button
-              className="btn-secondary mt-2 first-row__buttons--btn btn-cancel"
-              onClick={cancelEditProfile}
-            >
-              Cancel
-            </Button>
-            <Button
-              className="btn-primary mt-2 first-row__buttons--btn"
-              onClick={updateProfileFieldsDb}
-            >
-              Save
-            </Button>
+            {!isProfileInEditMode && (
+              <Row className="first-row__buttons">
+                {loadingRequest && (
+                  <Row>
+                    <Col colSpan={8} className="first-row__center-spinner">
+                      <Spinner animation="grow" className="bg-primary" />
+                    </Col>
+                  </Row>
+                )}
+                {errorMsg && (
+                  <div className="first-row__buttons--error error-message">
+                    {errorMsg}
+                  </div>
+                )}
+
+                <Button
+                  className="btn-secondary mt-2 first-row__buttons--btn btn-cancel"
+                  onClick={cancelEditProfile}
+                >
+                  Cancel
+                </Button>
+                <Button
+                  className="btn-primary mt-2 first-row__buttons--btn"
+                  onClick={updateProfileFieldsDb}
+                >
+                  Save
+                </Button>
+              </Row>
+            )}
           </Row>
-        )}
-      </Row>
-      <Row className="second-row">
-        <Col sm className="second-row__col">
-          <Col sm={4} className="second-row__col--flex-number primary-color">
-            35
-          </Col>
-          <Col sm={8}>
-            <Card.Text className="second-row__col--title primary-color">
-              Envato Website
-            </Card.Text>
-            <Card.Text>15 July</Card.Text>
-            <Card.Text>15 August</Card.Text>
-          </Col>
-        </Col>
-        <Col sm className="second-row__col">
-          <Col sm={4} className="second-row__col--flex-number success-color">
-            35
-          </Col>
-          <Col sm={8}>
-            <Card.Text className="second-row__col--title success-color">
-              ThemeForest CMS
-            </Card.Text>
-            <Card.Text>15 July</Card.Text>
-            <Card.Text>15 August</Card.Text>
-          </Col>
-        </Col>
-      </Row>
+          <Row className="second-row">
+            <Col sm className="second-row__col">
+              <Col
+                sm={4}
+                className="second-row__col--flex-number primary-color"
+              >
+                35
+              </Col>
+              <Col sm={8}>
+                <Card.Text className="second-row__col--title primary-color">
+                  Envato Website
+                </Card.Text>
+                <Card.Text>15 July</Card.Text>
+                <Card.Text>15 August</Card.Text>
+              </Col>
+            </Col>
+            <Col sm className="second-row__col">
+              <Col
+                sm={4}
+                className="second-row__col--flex-number success-color"
+              >
+                35
+              </Col>
+              <Col sm={8}>
+                <Card.Text className="second-row__col--title success-color">
+                  ThemeForest CMS
+                </Card.Text>
+                <Card.Text>15 July</Card.Text>
+                <Card.Text>15 August</Card.Text>
+              </Col>
+            </Col>
+          </Row>
+        </Container>
+      </Container>
     </Container>
   );
 };
